+++ node/network/bridge/Cargo.toml
@@ -19,6 +19,8 @@ polkadot-node-network-protocol = { path = "../protocol" }
 polkadot-node-subsystem-util = { path = "../../subsystem-util"}
 parking_lot = "0.12.0"
 bytes = "1"
+fatality = "0.0.6"
+thiserror = "1"
 
 [dev-dependencies]
 assert_matches = "1.4.0"
