+++ node/network/collator-protocol/src/validator_side/tests.rs
@@ -260,7 +260,7 @@ async fn assert_candidate_backing_second(
 async fn assert_collator_disconnect(virtual_overseer: &mut VirtualOverseer, expected_peer: PeerId) {
 	assert_matches!(
 		overseer_recv(virtual_overseer).await,
-		AllMessages::NetworkBridge(NetworkBridgeMessage::DisconnectPeer(
+		AllMessages::NetworkBridgeTx(NetworkBridgeTxMessage::DisconnectPeer(
 			peer,
 			peer_set,
 		)) => {
@@ -278,7 +278,7 @@ async fn assert_fetch_collation_request(
 ) -> ResponseSender {
 	assert_matches!(
 		overseer_recv(virtual_overseer).await,
-		AllMessages::NetworkBridge(NetworkBridgeMessage::SendRequests(reqs, IfDisconnected::ImmediateError)
+		AllMessages::NetworkBridgeTx(NetworkBridgeTxMessage::SendRequests(reqs, IfDisconnected::ImmediateError)
 	) => {
 		let req = reqs.into_iter().next()
 			.expect("There should be exactly one request");
@@ -431,8 +431,8 @@ fn collator_reporting_works() {
 
 		assert_matches!(
 			overseer_recv(&mut virtual_overseer).await,
-			AllMessages::NetworkBridge(
-				NetworkBridgeMessage::ReportPeer(peer, rep),
+			AllMessages::NetworkBridgeTx(
+				NetworkBridgeTxMessage::ReportPeer(peer, rep),
 			) => {
 				assert_eq!(peer, peer_b);
 				assert_eq!(rep, COST_REPORT_BAD);
@@ -481,8 +481,8 @@ fn collator_authentication_verification_works() {
 		// it should be reported for sending a message with an invalid signature
 		assert_matches!(
 			overseer_recv(&mut virtual_overseer).await,
-			AllMessages::NetworkBridge(
-				NetworkBridgeMessage::ReportPeer(peer, rep),
+			AllMessages::NetworkBridgeTx(
+				NetworkBridgeTxMessage::ReportPeer(peer, rep),
 			) => {
 				assert_eq!(peer, peer_b);
 				assert_eq!(rep, COST_INVALID_SIGNATURE);
@@ -697,7 +697,7 @@ fn reject_connection_to_next_group() {
 
 		assert_matches!(
 			overseer_recv(&mut virtual_overseer).await,
-			AllMessages::NetworkBridge(NetworkBridgeMessage::ReportPeer(
+			AllMessages::NetworkBridgeTx(NetworkBridgeTxMessage::ReportPeer(
 				peer,
 				rep,
 			)) => {
@@ -790,7 +790,7 @@ fn fetch_next_collation_on_invalid_collation() {
 
 		assert_matches!(
 			overseer_recv(&mut virtual_overseer).await,
-			AllMessages::NetworkBridge(NetworkBridgeMessage::ReportPeer(
+			AllMessages::NetworkBridgeTx(NetworkBridgeTxMessage::ReportPeer(
 				peer,
 				rep,
 			)) => {
@@ -1005,7 +1005,7 @@ fn disconnect_if_wrong_declare() {
 
 		assert_matches!(
 			overseer_recv(&mut virtual_overseer).await,
-			AllMessages::NetworkBridge(NetworkBridgeMessage::ReportPeer(
+			AllMessages::NetworkBridgeTx(NetworkBridgeTxMessage::ReportPeer(
 				peer,
 				rep,
 			)) => {
