+++ node/overseer/src/dummy.rs
@@ -85,6 +85,7 @@ pub fn dummy_overseer_builder<'a, Spawner, SupportsParachains>(
 		DummySubsystem,
 		DummySubsystem,
 		DummySubsystem,
+		DummySubsystem,
 	>,
 	SubsystemError,
 >
@@ -126,6 +127,7 @@ pub fn one_for_all_overseer_builder<'a, Spawner, SupportsParachains, Sub>(
 		Sub,
 		Sub,
 		Sub,
+		Sub,
 	>,
 	SubsystemError,
 >
@@ -143,7 +145,8 @@ where
 		+ Subsystem<OverseerSubsystemContext<ChainApiMessage>, SubsystemError>
 		+ Subsystem<OverseerSubsystemContext<CollationGenerationMessage>, SubsystemError>
 		+ Subsystem<OverseerSubsystemContext<CollatorProtocolMessage>, SubsystemError>
-		+ Subsystem<OverseerSubsystemContext<NetworkBridgeMessage>, SubsystemError>
+		+ Subsystem<OverseerSubsystemContext<NetworkBridgeRxMessage>, SubsystemError>
+		+ Subsystem<OverseerSubsystemContext<NetworkBridgeTxMessage>, SubsystemError>
 		+ Subsystem<OverseerSubsystemContext<ProvisionerMessage>, SubsystemError>
 		+ Subsystem<OverseerSubsystemContext<RuntimeApiMessage>, SubsystemError>
 		+ Subsystem<OverseerSubsystemContext<StatementDistributionMessage>, SubsystemError>
@@ -169,7 +172,8 @@ where
 		.chain_api(subsystem.clone())
 		.collation_generation(subsystem.clone())
 		.collator_protocol(subsystem.clone())
-		.network_bridge(subsystem.clone())
+		.network_bridge_tx(subsystem.clone())
+		.network_bridge_rx(subsystem.clone())
 		.provisioner(subsystem.clone())
 		.runtime_api(subsystem.clone())
 		.statement_distribution(subsystem.clone())
