+++ node/subsystem-types/src/messages.rs
@@ -318,9 +318,36 @@ pub enum DisputeDistributionMessage {
 	SendDispute(DisputeMessage),
 }
 
-/// Messages received by the network bridge subsystem.
+/// Messages received from other subsystems.
 #[derive(Debug)]
-pub enum NetworkBridgeMessage {
+pub enum NetworkBridgeRxMessage {
+	/// Inform the distribution subsystems about the new
+	/// gossip network topology formed.
+	///
+	/// The only reason to have this here, is the availability of the
+	/// authority discovery service, otherwise, the `GossipSupport`
+	/// subsystem would make more sense.
+	NewGossipTopology {
+		/// The session info this gossip topology is concerned with.
+		session: SessionIndex,
+		/// Ids of our neighbors in the X dimensions of the new gossip topology,
+		/// along with their validator indices within the session.
+		///
+		/// We're not necessarily connected to all of them, but we should
+		/// try to be.
+		our_neighbors_x: HashMap<AuthorityDiscoveryId, ValidatorIndex>,
+		/// Ids of our neighbors in the X dimensions of the new gossip topology,
+		/// along with their validator indices within the session.
+		///
+		/// We're not necessarily connected to all of them, but we should
+		/// try to be.
+		our_neighbors_y: HashMap<AuthorityDiscoveryId, ValidatorIndex>,
+	},
+}
+
+/// Messages received from other subsystems by the network bridge subsystem.
+#[derive(Debug)]
+pub enum NetworkBridgeTxMessage {
 	/// Report a peer for their actions.
 	ReportPeer(PeerId, UnifiedReputationChange),
 
@@ -375,27 +402,9 @@ pub enum NetworkBridgeMessage {
 		/// The peer set we want the connection on.
 		peer_set: PeerSet,
 	},
-	/// Inform the distribution subsystems about the new
-	/// gossip network topology formed.
-	NewGossipTopology {
-		/// The session info this gossip topology is concerned with.
-		session: SessionIndex,
-		/// Ids of our neighbors in the X dimensions of the new gossip topology,
-		/// along with their validator indices within the session.
-		///
-		/// We're not necessarily connected to all of them, but we should
-		/// try to be.
-		our_neighbors_x: HashMap<AuthorityDiscoveryId, ValidatorIndex>,
-		/// Ids of our neighbors in the X dimensions of the new gossip topology,
-		/// along with their validator indices within the session.
-		///
-		/// We're not necessarily connected to all of them, but we should
-		/// try to be.
-		our_neighbors_y: HashMap<AuthorityDiscoveryId, ValidatorIndex>,
-	},
 }
 
-impl NetworkBridgeMessage {
+impl NetworkBridgeTxMessage {
 	/// If the current variant contains the relay parent hash, return it.
 	pub fn relay_parent(&self) -> Option<Hash> {
 		match self {
@@ -408,7 +417,6 @@ impl NetworkBridgeMessage {
 			Self::ConnectToValidators { .. } => None,
 			Self::ConnectToResolvedValidators { .. } => None,
 			Self::SendRequests { .. } => None,
-			Self::NewGossipTopology { .. } => None,
 		}
 	}
 }
